frontend/requirements.txt
# Streamlit dependencies

streamlit
requests
pillow
